Much better buffer pooling strategy
Removes `crossbeam` dependency, buffers are instead pooled in a simple Vec,
because a `Server` only ever accesses the buffer pool from a single thread at a
time.

Acquiring a pooled buffer also no longer incurs the penalty of an Arc clone,
instead there is a borrowed handle form of a pooled buffer that borrows from an
instance of the shared pool and returns to the pool on drop.  There is still a
'static form of a buffer (`OwnedBuffer`) for moving messages and packets around,
but in order to access it it must be adopted by a `BufferPool` which discourages
pooled buffers from being lost during errors.  The result is instead of every
buffer having its own buffer pool handle , there is only the buffer pool for
each individaul client (and an extra one for the SSL stream because the openssl
API is not great).

We could use something faster than `Mutex::try_lock().unwrap()` in the future to
assert that the pool is only accessed from one thread at a time.