Remove unnecessary connection status distinction
The distinction between a client which has not yet fully connected and a client
which has disconnected is meaningless.  There is no way to *discover* clients
that have not yet fully connected, so the only clients that the user would be
aware of are fully connected ones.  Once a connected client begins the process
of shutting down, it is permanently "disconnected" from the perspective of the
outer API, thus there is only one error that needs to be handled: a connected
client becoming lost.