DTLS shutdown take.. 6? 7?
1. Actually send the shutdown packet immediately when the user calls
   Server::disconnect.  Otherwise, it will sit there until something else causes
   packets to be sent for that client, which can be a long time.  Makes
   disconnect into an async method.

2. Make sure to allow packets to be sent out if shutdown returns 'received',
   because it may have now sent a response packet.

3. Handle ZERO_RETURN from shutdown as meaning that the connection is completely
   shut down and no further calls to shutdown are necessary.

This should solve DTLS shutdown at least in the case that the shutdown packet is
not lost.  I do not know how shutdown is supposed to work in the presence of
packet loss, SSL_shutdown is not very specific on how it's supposed to work for
DTLS, so rather than guess I am not going to attempt to handle it.